feat: Remove the common crate and move definitions to scheduler crate
