+++ client/src/lib.rs
@@ -4,14 +4,15 @@ use std::time::Duration;
 
 use tracing::{debug, error, trace, warn};
 
-pub use common::{
+pub use error::Error;
+pub use rpc_client::RpcCaller;
+use scheduler::run_scheduler;
+pub use scheduler::{
     list_devices, ClientToken, Deadline, DeviceId, Devices, Pid, PreemptionResponse, ResourceAlloc,
     ResourceMemory, ResourceReq, ResourceType, TaskEstimations, TaskFunc, TaskReqBuilder,
     TaskRequirements, TaskResult, TaskType,
 };
-pub use error::Error;
-pub use rpc_client::RpcCaller;
-use scheduler::run_scheduler;
+
 pub use scheduler::{spawn_scheduler_with_handler, Error as SchedulerError, Settings};
 use std::path::PathBuf;
 
@@ -311,7 +312,7 @@ impl Client {
         use nix::unistd::{fork, ForkResult};
 
         // check if there are resources to manage before trying to start the scheduler service
-        let devices = common::list_devices();
+        let devices = scheduler::list_devices();
         if devices.gpu_devices().is_empty() {
             return Err(Error::NoGpuResources);
         }
@@ -416,7 +417,7 @@ mod tests {
         let mut settings = Settings::new(path).unwrap();
         let socket = UdpSocket::bind("127.0.0.1:0").unwrap();
         settings.service.address = format!("{}", socket.local_addr().unwrap());
-        let devices = common::list_devices();
+        let devices = scheduler::list_devices();
         let handle =
             spawn_scheduler_with_handler(settings.clone(), "/tmp/one_process/", devices).unwrap();
         let client = Client::register_with_settings::<Error>(settings).unwrap();
@@ -437,7 +438,7 @@ mod tests {
         let mut settings = Settings::new(path).unwrap();
         let socket = UdpSocket::bind("127.0.0.1:0").unwrap();
         settings.service.address = format!("{}", socket.local_addr().unwrap());
-        let devices = common::list_devices();
+        let devices = scheduler::list_devices();
         let handle =
             spawn_scheduler_with_handler(settings.clone(), "/tmp/release_test/", devices).unwrap();
         let client = Client::register_with_settings::<Error>(settings).unwrap();
@@ -455,7 +456,7 @@ mod tests {
         let mut settings = Settings::new(path).unwrap();
         let socket = UdpSocket::bind("127.0.0.1:0").unwrap();
         settings.service.address = format!("{}", socket.local_addr().unwrap());
-        let devices = common::list_devices();
+        let devices = scheduler::list_devices();
         let handle =
             spawn_scheduler_with_handler(settings.clone(), "/tmp/panic_test/", devices).unwrap();
         let client = Client::register_with_settings::<Error>(settings).unwrap();
