--- common/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-authors = ["neithanmo <neithanmo@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-chrono = { version = "0.4.19", features = ["serde"] }
-num_cpus = "1.13.0"
-rust-gpu-tools = { version = "0.4.0"}
-serde = { version = "1.0.120", features = ["serde_derive"] }
-tracing = "0.1.25"
-palaver = "0.2.8"
