+++ scheduler/src/requests.rs
@@ -1,10 +1,28 @@
 use futures::channel::oneshot;
 use serde::{Deserialize, Serialize};
 
-use common::{DeviceId, Pid, PreemptionResponse, RequestMethod, ResourceAlloc};
-
 use crate::monitor::MonitorInfo;
-use crate::Result;
+use crate::{ClientToken, DeviceId, Pid, ResourceAlloc, Result, TaskRequirements};
+
+#[derive(Serialize, Deserialize)]
+pub enum RequestMethod {
+    Schedule(ClientToken, TaskRequirements, String),
+    ListAllocations,
+    WaitPreemptive(ClientToken),
+    Release(ClientToken),
+    ReleasePreemptive(ClientToken),
+    Abort(Vec<Pid>),
+    RemoveStalled(Vec<Pid>),
+    CheckService,
+    Monitoring,
+}
+
+#[derive(Serialize, Deserialize, Eq, PartialEq, Debug)]
+pub enum PreemptionResponse {
+    Execute,
+    Wait,
+    Abort,
+}
 
 #[derive(Serialize, Deserialize)]
 pub enum SchedulerResponse {
